requirements.txt
appdirs==1.4.4
astroid==2.5.6
attrs==21.2.0
backcall==0.2.0
black==21.5b1
click==8.0.1
colorama==0.4.4
coverage==5.5
cycler==0.10.0
decorator==5.0.9
docopt==0.6.2
gitdb==4.0.7
GitPython==3.1.17
gpytorch==1.4.2
iniconfig==1.1.1
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
joblib==1.0.1
jsonpickle==1.5.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
progressbar==2.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
py-cpuinfo==8.0.0
Pygments==2.9.0
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
regex==2021.4.4
sacred==0.8.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
smmap==4.0.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1+cu111
tqdm==4.61.0
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
wrapt==1.12.1
